improved projects lookup, including autocomplete triggers (#223)
