+++ src/index.ts
@@ -6,5 +6,5 @@ export {
   ReferenceFunctionParams,
   ExtensionCapabilities,
 } from './utils/addon-api';
-export { Project, Executor, Destructor, Linter, Watcher } from './project-roots';
+export { Project, Executor, Destructor, Linter, Watcher } from './project';
 export { default as Server } from './server';
