+++ src/template-linter.ts
@@ -11,7 +11,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 
 import Server from './server';
-import { Project } from './project-roots';
+import { Project } from './project';
 
 export interface TemplateLinterError {
   fatal?: boolean;
