+++ src/utils/addon-api.ts
@@ -17,7 +17,7 @@ import CoreScriptDefinitionProvider from './../builtin-addons/core/script-defini
 import CoreTemplateDefinitionProvider from './../builtin-addons/core/template-definition-provider';
 import ScriptCompletionProvider from './../builtin-addons/core/script-completion-provider';
 import TemplateCompletionProvider from './../builtin-addons/core/template-completion-provider';
-import { Project } from '../project-roots';
+import { Project } from '../project';
 
 import TemplateLintFixesCodeAction from '../builtin-addons/core/code-actions/template-lint-fixes';
 import TemplateLintCommentsCodeAction from '../builtin-addons/core/code-actions/template-lint-comments';
